+++ liquibase-standard/src/main/java/liquibase/parser/core/formattedsql/FormattedSqlChangeLogParser.java
@@ -7,6 +7,7 @@ import liquibase.changelog.ChangeSet;
 import liquibase.exception.ChangeLogParseException;
 import liquibase.parser.FormattedChangeLogParser;
 import liquibase.precondition.core.PreconditionContainer;
+import liquibase.precondition.core.SqlPrecondition;
 import liquibase.util.StringUtil;
 
 import java.util.regex.Matcher;
@@ -20,6 +21,16 @@ public class FormattedSqlChangeLogParser extends FormattedChangeLogParser {
 
     private static final String ON_SQL_OUTPUT_REGEX = ".*onSqlOutput:(\\w+).*";
     private static final Pattern ON_SQL_OUTPUT_PATTERN = Pattern.compile(ON_SQL_OUTPUT_REGEX, Pattern.CASE_INSENSITIVE);
+
+    private static final String WORD_RESULT_REGEX = "^(?:expectedResult:)?(\\w+) (.*)";
+    private static final String SINGLE_QUOTE_RESULT_REGEX = "^(?:expectedResult:)?'([^']+)' (.*)";
+    private static final String DOUBLE_QUOTE_RESULT_REGEX = "^(?:expectedResult:)?\"([^\"]+)\" (.*)";
+
+    private static final Pattern[] WORD_AND_QUOTING_PATTERNS = new Pattern[]{
+            Pattern.compile(WORD_RESULT_REGEX, Pattern.CASE_INSENSITIVE),
+            Pattern.compile(SINGLE_QUOTE_RESULT_REGEX, Pattern.CASE_INSENSITIVE),
+            Pattern.compile(DOUBLE_QUOTE_RESULT_REGEX, Pattern.CASE_INSENSITIVE)
+    };
 
     @Override
     protected String getStartMultiLineCommentSequence() {
@@ -69,6 +80,38 @@ public class FormattedSqlChangeLogParser extends FormattedChangeLogParser {
             }
             changeSet.setPreconditions(pc);
         }
+    }
+
+    @Override
+    protected void handlePreconditionCase(ChangeLogParameters changeLogParameters, ChangeSet changeSet, Matcher preconditionMatcher) throws ChangeLogParseException {
+        if (changeSet.getPreconditions() == null) {
+            // create the defaults
+            changeSet.setPreconditions(new PreconditionContainer());
+        }
+        if (preconditionMatcher.groupCount() == 2) {
+            String name = StringUtil.trimToNull(preconditionMatcher.group(1));
+            if (name != null) {
+                String body = preconditionMatcher.group(2).trim();
+                if ("sql-check".equals(name)) {
+                    changeSet.getPreconditions().addNestedPrecondition(parseSqlCheckCondition(changeLogParameters.expandExpressions(StringUtil.trimToNull(body), changeSet.getChangeLog())));
+                } else {
+                    throw new ChangeLogParseException("The '" + name + "' precondition type is not supported.");
+                }
+            }
+        }
+    }
+
+    private SqlPrecondition parseSqlCheckCondition(String body) throws ChangeLogParseException{
+        for (Pattern pattern : WORD_AND_QUOTING_PATTERNS) {
+            Matcher matcher = pattern.matcher(body);
+            if (matcher.matches() && (matcher.groupCount() == 2)) {
+                SqlPrecondition p = new SqlPrecondition();
+                p.setExpectedResult(matcher.group(1));
+                p.setSql(matcher.group(2));
+                return p;
+            }
+        }
+        throw new ChangeLogParseException("Could not parse a SqlCheck precondition from '" + body + "'.");
     }
 
     @Override
